pyproject.toml
[tool.poetry]
name = "jaemon"
version = "0.1.0"
description = "Background job alert service"
authors = ["Daniel Schirmacher <dschirethz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "*"
beautifulsoup4 = "*"
apscheduler = "*"
PyYAML = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
pre-commit = "*"
ruff = "*"
black = "*"
isort = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
